add usage client to clientset
