+++ client/azure_client_set.go
@@ -1,8 +1,6 @@
 package client
 
 import (
-	"errors"
-
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2018-06-01/compute"
 	"github.com/Azure/azure-sdk-for-go/services/dns/mgmt/2017-10-01/dns"
 	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2018-06-01/network"
@@ -13,44 +11,6 @@ import (
 	"github.com/Azure/go-autorest/autorest/azure"
 	"github.com/giantswarm/microerror"
 )
-
-type AzureClientSetConfig struct {
-	// ClientID is the ID of the Active Directory Service Principal.
-	ClientID string
-	// ClientSecret is the secret of the Active Directory Service Principal.
-	ClientSecret string
-	// The cloud environment identifier. Takes values from
-	// https://github.com/Azure/go-autorest/blob/ec5f4903f77ed9927ac95b19ab8e44ada64c1356/autorest/azure/environments.go#L13
-	EnvironmentName string
-	// SubscriptionID is the ID of the Azure subscription.
-	SubscriptionID string
-	// TenantID is the ID of the Active Directory tenant.
-	TenantID string
-}
-
-// clientConfig contains all essential information to create an Azure client.
-type clientConfig struct {
-	subscriptionID          string
-	resourceManagerEndpoint string
-	servicePrincipalToken   *adal.ServicePrincipalToken
-}
-
-func (c AzureClientSetConfig) Validate() error {
-	if c.ClientID == "" {
-		return errors.New("ClientID must not be empty")
-	}
-	if c.ClientSecret == "" {
-		return errors.New("ClientSecret must not be empty")
-	}
-	if c.SubscriptionID == "" {
-		return errors.New("SubscriptionID must not be empty")
-	}
-	if c.TenantID == "" {
-		return errors.New("TenantID must not be empty")
-	}
-
-	return nil
-}
 
 // AzureClientSet is the collection of Azure API clients.
 type AzureClientSet struct {
@@ -66,6 +26,8 @@ type AzureClientSet struct {
 	InterfacesClient *network.InterfacesClient
 	//StorageAccountsClient manages blobs in storage containers.
 	StorageAccountsClient *storage.AccountsClient
+	// UsageClient is used to work with limits and quotas.
+	UsageClient *compute.UsageClient
 	// VirtualNetworkClient manages virtual networks.
 	VirtualNetworkClient *network.VirtualNetworksClient
 	// VirtualNetworkGatewayConnectionsClient manages virtual network gateway connections.
@@ -82,20 +44,21 @@ type AzureClientSet struct {
 
 // NewAzureClientSet returns the Azure API clients.
 func NewAzureClientSet(config AzureClientSetConfig) (*AzureClientSet, error) {
-	if err := config.Validate(); err != nil {
-		return nil, microerror.Maskf(invalidConfigError, "config.%s", err)
+	err := config.Validate()
+	if err != nil {
+		return nil, microerror.Mask(err)
 	}
 
 	// Returns environment object contains all API endpoints for specific Azure
 	// cloud. For empty config.EnvironmentName returns Azure public cloud.
 	env, err := parseAzureEnvironment(config.EnvironmentName)
 	if err != nil {
-		return nil, microerror.Maskf(err, "parsing Azure environment")
+		return nil, microerror.Mask(err)
 	}
 
 	servicePrincipalToken, err := newServicePrincipalToken(config, env)
 	if err != nil {
-		return nil, microerror.Maskf(err, "creating service principal token")
+		return nil, microerror.Mask(err)
 	}
 
 	c := &clientConfig{
@@ -106,11 +69,12 @@ func NewAzureClientSet(config AzureClientSetConfig) (*AzureClientSet, error) {
 
 	clientSet := &AzureClientSet{
 		DeploymentsClient:                      newDeploymentsClient(c),
-		GroupsClient:                           newGroupsClient(c),
 		DNSRecordSetsClient:                    newDNSRecordSetsClient(c),
 		DNSZonesClient:                         newDNSZonesClient(c),
+		GroupsClient:                           newGroupsClient(c),
 		InterfacesClient:                       newInterfacesClient(c),
 		StorageAccountsClient:                  newStorageAccountsClient(c),
+		UsageClient:                            newUsageClient(c),
 		VirtualNetworkClient:                   newVirtualNetworkClient(c),
 		VirtualNetworkGatewayConnectionsClient: newVirtualNetworkGatewayConnectionsClient(c),
 		VirtualNetworkGatewaysClient:           newVirtualNetworkGatewaysClient(c),
@@ -160,6 +124,13 @@ func newInterfacesClient(config *clientConfig) *network.InterfacesClient {
 func newStorageAccountsClient(config *clientConfig) *storage.AccountsClient {
 	c := storage.NewAccountsClient(config.subscriptionID)
 	c.Authorizer = autorest.NewBearerAuthorizer(config.servicePrincipalToken)
+	return &c
+}
+
+func newUsageClient(config *clientConfig) *compute.UsageClient {
+	c := compute.NewUsageClientWithBaseURI(config.resourceManagerEndpoint, config.subscriptionID)
+	c.Authorizer = autorest.NewBearerAuthorizer(config.servicePrincipalToken)
+
 	return &c
 }
